Moved UserAuth-only components from main components directory to UserAuth page directory.
